blog-generator-be/pyproject.toml
[project]
name = "blog-generator-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "langchain-openai>=0.3.24",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.8",
    "langgraph-cli[inmem]>=0.3.3",
    "uvicorn>=0.34.3",
    "watchdog>=6.0.0",
    "python-dotenv>=1.1.0",
    "openai>=1.88.0",
]
